Refactored appsec coverage
