+++ neo4j_appsec_coverage_loader/pipeline.py
@@ -3,4 +3,4 @@ from etwpipeline.declarative import DeclarativePipeline
 
 
 def make_pipeline() -> Pipeline:
-    return DeclarativePipeline.from_file("neo4j_appsec_loader/akamai-appsec.yaml")
+    return DeclarativePipeline.from_file("neo4j_appsec_coverage_loader/akamai-appsec-coverage.yaml")
