+++ neo4j_waf_loader/akamai-waf.yaml
@@ -20,9 +20,9 @@ transformers:
           status: !!python/jmespath status
         - type: relationship
           iterate_on: !!python/jmespath policies[*]
-          search: !!python/jmespath policyName
+          search: !!python/jmespath policyId
           related_node_type: akamai_waf_policy
-          related_field_name: policyName
+          related_field_name: policyId
           relationship_type: DEFINED_BY
